requirements.txt
genanki
requests_html
lxml_html_clean
PyQt6
pytest-qt